+++ basic_feedforward_nn/inc/utils/verity_results.h
@@ -4,8 +4,8 @@
 #include <vector>
 #include <cassert>
 
-void verify_result_mul(std::vector<int>& a, std::vector<int>& b, std::vector<int>& c, int N);
-void verify_result_add(std::vector<int>& a, std::vector<int>& b, std::vector<int>& c, int N);
-void verify_result_scale(std::vector<int>& a, int factor, std::vector<int>& c, int N);
+void verify_result_mul(std::vector<float>& a, std::vector<float>& b, std::vector<float>& c, int N);
+void verify_result_add(std::vector<float>& a, std::vector<float>& b, std::vector<float>& c, int N);
+void verify_result_scale(std::vector<float>& a, int factor, std::vector<float>& c, int N);
 
 #endif // !UTILS_VERIFY_RESULTS_H
